+++ java-design-processor/src/test/resources/internal/nbbrd/design/TestDecoratorConstructorNoDecorated.java
@@ -1,0 +1,11 @@
+package internal.nbbrd.design;
+
+import nbbrd.design.DecoratorPattern;
+
+import java.io.Reader;
+
+@DecoratorPattern
+public abstract class TestDecoratorConstructorNoDecorated extends Reader {
+
+    TestDecoratorConstructorNoDecorated(boolean stuff) {}
+}