+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2025-05-08
+
+### Added
+
+- Add DecoratorPattern annotation [#338](https://github.com/nbbrd/java-design-util/issues/338)
+- Add support of external factories in StaticFactoryMethod [#152](https://github.com/nbbrd/java-design-util/issues/152)
+
+### Fixed
+
+- Fix package pattern in MightBePromoted [#337](https://github.com/nbbrd/java-design-util/issues/337)
+
 ## [1.5.1] - 2024-03-27
 
 ### Fixed
@@ -98,7 +109,8 @@ This release replaces the `StringValue` annotation with a more general approach 
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-design-util/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/nbbrd/java-design-util/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/nbbrd/java-design-util/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/nbbrd/java-design-util/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/nbbrd/java-design-util/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/nbbrd/java-design-util/compare/v1.3.1...v1.4.0